django_blog/requirements.txt
Django>=4.2,<6.0
django-taggit>=5.0
